+++ src/main/java/master/innutrient/client/gui/DashboardTooltipSource.java
@@ -1,0 +1,9 @@
+package master.innutrient.client.gui;
+
+import net.minecraft.network.chat.Component;
+
+import java.util.List;
+
+interface DashboardTooltipSource {
+    List<Component> tooltipAt(double mouseX, double mouseY);
+}